updated bg-color/padding on knowledgebase link #133 (#145)
* updated bg-color/padding on knowledgebase link

* Get rid of '.link-pinned.ungray.top' classes & div.

Less is more:) So this is a great spot to use the natural spacing provided by our framework extensions. Stacking rows adds margin-top to all rows AFTER the first row. The '.row-short' class uses 15px of margin-top rather than the default 30px.

* Removing '.ungray' class

We don't need to unstyle a class, rather remove it completely.

* Added translation to kbase <a> tag