Put the commit hash in the rev query param of the JS and CSS includes
Before this patch there was some gulp task that set those two includes, but another release process added the commit hash elsewhere.
There was also some bug where the rev= gulp task thing would not change on production-stable. It worked on master though.

This brings some consistency because the old hash was not the commit has, I think it was a sum of the files themselves or something.